+++ src/main/java/de/ellpeck/game/world/tile/Tile.java
@@ -172,13 +172,13 @@ public class Tile{
     public void doPlace(World world, int x, int y, TileLayer layer, ItemInstance instance, EntityPlayer placer){
         world.setTile(layer, x, y, this);
 
-        byte meta = this.getPlacementMeta(world, x, y, layer, instance);
+        int meta = this.getPlacementMeta(world, x, y, layer, instance);
         if(meta != 0){
             world.setMeta(layer, x, y, meta);
         }
     }
 
-    public byte getPlacementMeta(World world, int x, int y, TileLayer layer, ItemInstance instance){
+    public int getPlacementMeta(World world, int x, int y, TileLayer layer, ItemInstance instance){
         return instance.getMeta();
     }
